oldapi/requirements.txt
email-validator==2.2.0
Flask-SQLAlchemy==3.0.2
python-dateutil==2.9.0
requests==2.28.1
sqlalchemy-utils==0.38.3
logtail-python==0.3.0